Apply @utils.measure
